requirements.txt
flask>=2.3.3
requests>=2.31.0
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.25.2
matplotlib>=3.7.2
streamlit>=1.25.0
pytest>=7.4.0
python-dotenv>=1.0.0
altair>=5.0.1
psycopg2-binary>=2.9.6 